# Optimize db_stress setup phase